Add entry for mock in sudoers
Mock needs to run with root permissions but not directly with the root
user. In Clear Linux the way to run it is using sudo. This patch adds
permissions for whoever is in the mock group to be able to use sudo to
run mock.

For reference:

- In the past mock had a C compiled wrapper to call mock itself, and
made the wrapper have the suidbit (with root as owner). The wrapper was
removed from mock.

- Currently in Fedora, consolehelper is used to provide the necessary
privileges, it is a generalization from the mechanism above.